+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file
 - add b9partswitch and cryotanks compat
 - add KAS compat
 - patch Kethane to use server-wide seed
+- patch ExtraplanetaryLaunchpad recylcer to use a server-wide random
 
 
 ## 0.0.1 - 2025-06-27